requirements.txt
django>=1.10,<1.11
django-sortedm2m>=1.3.2<1.4